backend/requirements.txt
fastapi>=0.116,<1.0
uvicorn[standard]>=0.35,<1.0
pydantic>=2.11,<3.0
requests>=2.32,<3.0
pandas>=2.3,<3.0
sqlalchemy>=2.0,<3.0
psycopg[binary]>=3.2,<4.0
alembic>=1.16,<2.0
PyJWT>=2.10,<3.0
passlib>=1.7,<2.0
email-validator>=2.2,<3.0
